Moved templatetags tests out of templatetags module to fix build
