+++ main/templatetags_test.py
@@ -1,12 +1,17 @@
 """
-Tests for render_bundle
+Tests for templatetags in the main Django app
 """
-from django.test.client import RequestFactory
+# NOTE: This file is located here and not in main/templatetags/ because Django attempts to load
+# all files in the 'templatetags' directory as template tags, even '_test.py' files.
 import pytest
+from datetime import datetime
+from pytz import utc
+from django.test.client import RequestFactory
 
 from main.utils import webpack_dev_server_url
 from main.templatetags.render_bundle import render_bundle, public_path
-
+from main.templatetags.dollar_format import dollar_format
+from main.templatetags.parse_date import parse_iso_datetime
 
 FAKE_COMMON_BUNDLE = [
     {
@@ -93,3 +98,18 @@ def test_missing_file(mocker):
 
     get_bundle.assert_called_with(bundle_name)
     get_loader.assert_called_with('DEFAULT')
+
+
+def test_dollar_format():
+    """Test that dollar_format takes a number representing a dollar value and formats it correctly"""
+    assert dollar_format('123') == '$123.00'
+    assert dollar_format('123.5') == '$123.50'
+    assert dollar_format(123) == '$123.00'
+    assert dollar_format(123.5) == '$123.50'
+
+
+def test_parse_iso_datetime():
+    """Test that parse_iso_datetime correctly formats an ISO 8601 date string"""
+    parsed_datetime = parse_iso_datetime('2017-01-01T01:01:01.000000Z')
+    assert parsed_datetime == datetime(year=2017, month=1, day=1, hour=1, minute=1, second=1, tzinfo=utc)
+    assert parse_iso_datetime('invalid') is None
